feat: diamond
